# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for sorting wrong problems.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "1.116.126.208:8080",
    "paths": {
...
        "/app/exercise/insert_exercise": {
            "post": {
                "description": "插入今日习题数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teacher"
                ],
                "summary": "插入今日习题",
                "parameters": [
                    {
                        "description": "习题数据",
                        "name": "ExerciseData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Models.Exercise"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\":200,\"msg\":\"Insert() success!\",\"data\":500001}",
                        "schema": {
                            "$ref": "#/definitions/Controllers.Res"
                        }
                    }
                }
            }
        },
...
        "/app/teacher/get_student": {
            "get": {
                "description": "教师获取监管校区内的一个学生信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teacher"
                ],
                "summary": "教师获取一个学生信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "学生id",
                        "name": "Sid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Controllers.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/Models.Student"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/app/teacher/get_teacher": {
            "get": {
                "description": "获取一个教师信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "获取一个教师信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "教师id",
                        "name": "Tid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Controllers.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/Services.Teacher"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/app/teacher/update_password": {
            "put": {
                "description": "编辑一个教师信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teacher"
                ],
                "summary": "编辑教师",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "name": "cnames",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "ticon",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "tid",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tnickname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "toldPassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tpassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tpasswordAgain",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tphone",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Controllers.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/Services.Teacher"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/app/teacher/update_teacher": {
            "post": {
                "description": "编辑一个教师信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "编辑教师",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "name": "cnames",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "ticon",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "tid",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tnickname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "toldPassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tpassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tpasswordAgain",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "tphone",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Controllers.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/Services.Teacher"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Controllers.Res": {
            "type": "object",
            "properties": {
                "Code": {
                    "type": "integer"
                },
                "Data": {
                    "type": "object"
                },
                "Msg": {
                    "type": "string"
                }
            }
        },
...
        "Models.Exercise": {
            "type": "object",
            "properties": {
                "eanswerPath": {
                    "type": "string"
                },
                "eanswerTxt": {
                    "type": "string"
                },
                "edescription": {
                    "type": "string"
                },
                "edifficulty": {
                    "type": "string"
                },
                "egrade": {
                    "type": "string"
                },
                "eid": {
                    "type": "integer"
                },
                "enamePath": {
                    "type": "string"
                },
                "enameTxt": {
                    "type": "string"
                },
                "esubject": {
                    "type": "string"
                },
                "etitle": {
                    "type": "string"
                },
                "etype": {
                    "type": "string"
                },
                "eunit": {
                    "type": "string"
                }
            }
        },
...
        "Models.Student": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer"
                },
                "sgrade": {
                    "type": "integer"
                },
                "sicon": {
                    "type": "string"
                },
                "sid": {
                    "type": "integer"
                },
                "sname": {
                    "type": "string"
                },
                "snickname": {
                    "type": "string"
                },
                "spassword": {
                    "type": "string"
                },
                "sphone": {
                    "type": "string"
                }
            }
        },
...
        "Services.Teacher": {
            "type": "object",
            "properties": {
                "cnames": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ticon": {
                    "type": "string"
                },
                "tid": {
                    "type": "integer"
                },
                "tname": {
                    "type": "string"
                },
                "tnickname": {
                    "type": "string"
                },
                "toldPassword": {
                    "type": "string"
                },
                "tpassword": {
                    "type": "string"
                },
                "tpasswordAgain": {
                    "type": "string"
                },
                "tphone": {
                    "type": "string"
                }
            }
        }
    }
}